move params client to module
