+++ x/params/client/proposal_handler.go
@@ -1,9 +1,9 @@
 package client
 
 import (
-	"github.com/cosmos/cosmos-sdk/params/client/cli"
-	"github.com/cosmos/cosmos-sdk/params/client/rest"
 	govclient "github.com/cosmos/cosmos-sdk/x/gov/client"
+	"github.com/cosmos/cosmos-sdk/x/params/client/cli"
+	"github.com/cosmos/cosmos-sdk/x/params/client/rest"
 )
 
 // param change proposal handler
